Changes for user session and bank account flow
