+++ src/pages/cards/components/WithdrawFundsPopup.tsx
@@ -1,8 +1,11 @@
-import Rect, { useState, useEffect } from "react";
+import Rect, { useState, useEffect, useContext } from "react";
 import { Form, Formik, FormikErrors, FormikProps } from "formik";
 import Spinner from "./Spinner";
 import { useNavigate } from "react-router-dom";
 import cardsAPIService from "../../../services/cardsAPIService";
+import { sessionContext } from "../../../app";
+import flareDBService from "../../../services/flareDBService";
+import Bank from "./bank-accounts/Bank";
 
 interface WithdrawData {
 	account_id: string;
@@ -74,10 +77,12 @@ interface WithdrawFundsPopupProps {
 
 const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 	const cardsService = new cardsAPIService();
+	const flareService = new flareDBService();
 	let navigate = useNavigate();
 	const [showSpinner, setShowSpinner] = useState(false);
 	const [message, setMessage] = useState("");
-	const account_id = sessionStorage.getItem("account_id");
+	const session = useContext(sessionContext);
+	const account_id = session?.organization?.stripeConnectId;
 	const initialValues: WithdrawData = {
 		account_id: account_id || "",
 		source_id: "src_1LNBMbR1aSxGwRclfo8FZZdU",
@@ -94,9 +99,10 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 	useEffect(() => {
 		if (isShow) {
 			retrieveBalance();
-			getAllSources();
+			getAllBankccounts();
 		}
 	}, [isShow]);
+
 	const retrieveBalance = async () => {
 		setShowSpinner(true);
 		try {
@@ -114,37 +120,31 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 			setShowSpinner(false);
 		}
 	};
-	const getAllSources = async () => {
-		setShowSpinner(true);
-		// try {
-		// 	let response: any = await cardsService.retrieveBalance({ account_id: account_id });
-		// 	setShowSpinner(false);
-		// 	if (
-		// 		response.type === "StripePermissionError" ||
-		// 		response.type === "StripeInvalidRequestError"
-		// 	) {
-		// 		//setMessage(data.raw.message);
-		// 	} else {
-		// 		setSources(response);
-		// 	}
-		// } catch (ex) {
-		// 	console.log("exception", ex);
-		// 	setShowSpinner(false);
-		// }
-		setSources([
-			{
-				id: "src_1LNBMbR1aSxGwRclfo8FZZdU",
-				object: "source",
-				ach_debit: {
-					last4: "9991",
-					bank_name: "STRIPE TEST BANK",
-				},
-				owner: {
-					name: "Jenny Rosen",
-				},
-			},
-		]);
+
+	const getAllBankccounts = async (): Promise<void> => {
+		const organization = session?.organization;
+		if (organization) {
+			setShowSpinner(true);
+			const organizationId = organization.id ? organization.id : "";
+			try {
+				let response: any = await flareService.getAllBankAccounts(organizationId);
+				setShowSpinner(false);
+				if (
+					response.type === "StripePermissionError" ||
+					response.type === "StripeInvalidRequestError"
+				) {
+					setMessage(response.raw.message);
+				} else {
+					setMessage("");
+					setSources(response);
+				}
+			} catch (ex) {
+				console.log("exception", ex);
+				setShowSpinner(false);
+			}
+		}
 	};
+
 	const onSubmit = (values: WithdrawData) => {
 		withdrawFunds(values);
 	};
@@ -167,7 +167,6 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 		try {
 			let response: any = await cardsService.withdrawFunds(values);
 			setShowSpinner(false);
-			debugger;
 			if (
 				response.type === "StripePermissionError" ||
 				response.type === "StripeInvalidRequestError"
@@ -188,12 +187,12 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 				<Formik
 					initialValues={initialValues}
 					validate={validate}
-					onSubmit={(values, actions) => {
+					onSubmit={(values: WithdrawData, actions: any) => {
 						onSubmit(values);
 						actions.setSubmitting(false);
 					}}
 				>
-					{(formik) => {
+					{(formik: FormikProps<WithdrawData>) => {
 						const { handleSubmit, handleChange, touched, errors, handleBlur, values } =
 							formik;
 
@@ -215,9 +214,9 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 							<>
 								<Spinner show={showSpinner} />
 								<Form className="connect-bk-form" onSubmit={handleSubmit}>
-									<div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full">
+									<div className="fixed inset-0 bg-gray-600 bg-opacity-50 overflow-y-auto h-full w-full z-30">
 										<div className="flex justify-center overflow-x-hidden overflow-y-auto px-6 py-6">
-											<div className="flex bg-white w-[817px] h-[500px] border-gray-300 rounded-md shadow-lg">
+											<div className="flex bg-white w-[817px] h-[525px] border-gray-300 rounded-md shadow-lg">
 												<div className="relative w-full h-full max-w-auto p-5">
 													<div className="flex justify-between items-center">
 														<p className="text-2xl text-gray-900 font-bold">
@@ -307,44 +306,38 @@ const WithdrawFundsPopup = ({ isShow, onHide }: WithdrawFundsPopupProps) => {
 															</div>
 														</div>
 													</div>
-													<div className="mt-10">
+													<div className="mt-5">
 														<div className="text-black-600 text-xs">
 															Withdraw Funds to
 														</div>
-														{sources.map((source: any) => (
-															<>
-																<div className=" text-black-500 text-sm">
-																	<input
-																		type="radio"
-																		name="source_id"
-																		className="mr-2 accent-red-600"
-																		onChange={handleChange}
-																		onBlur={handleBlur}
-																		value={source.id}
-																	/>
-																	{`${source.owner!.name} | ${
-																		source.ach_debit!.bank_name
-																	}`}
-																</div>
-																<div className=" text-gray-600 text-sm">
-																	************
-																	{source.ach_debit!.last4}
-																</div>
-															</>
-														))}
+														<div className="overflow-y-auto h-20">
+															{sources.map((source: any) => (
+																<Bank
+																	bank={source}
+																	handleChange={handleChange}
+																	handleBlur={handleBlur}
+																	key={source.id}
+																/>
+															))}
+															{isTouched && error && (
+																<p className="mt-2 text-sm text-red-600 dark:text-red-500">
+																	Select source from source list
+																</p>
+															)}
+														</div>
 														{isTouched && error && (
 															<p className="mt-2 text-sm text-red-600 dark:text-red-500">
 																Select source from source list
 															</p>
 														)}
-														<div
+														<a
 															className="text-sm text-red-500 mt-3 cursor-pointer hover:font-bold"
 															onClick={() => {
 																navigate("/connect-bank-account");
 															}}
 														>
 															+ Add Another Bank Account
-														</div>
+														</a>
 													</div>
 													<div className="flex justify-between space-x-4 mb-0 mt-6 border-t-2 pt-6">
 														<div className="text-sm font-bold ml-4 h-auto w-72">
