+++ src/pages/cards/components/bank-accounts/BankAccountListItem.tsx
@@ -1,0 +1,80 @@
+const BankAccountListItem = ({ bank, handleChange, handleBlur }: any) => {
+	return (
+		<>
+			<div className="flex justify-between p-2 border">
+				<div className="flex ">
+					<svg
+						width="26"
+						height="26"
+						viewBox="0 0 26 26"
+						fill="none"
+						xmlns="http://www.w3.org/2000/svg"
+					>
+						<path
+							d="M24.9376 11.4379C25.9032 11.4379 26.3064 10.1973 25.522 9.62539L13.5876 0.94414C13.4172 0.819271 13.2114 0.751953 13.0001 0.751953C12.7888 0.751953 12.583 0.819271 12.4126 0.94414L0.478237 9.62539C-0.306138 10.1941 0.0969874 11.4379 1.06574 11.4379H3.00011V23.1254H0.750112C0.612612 23.1254 0.500112 23.2379 0.500112 23.3754V25.0004C0.500112 25.1379 0.612612 25.2504 0.750112 25.2504H25.2501C25.3876 25.2504 25.5001 25.1379 25.5001 25.0004V23.3754C25.5001 23.2379 25.3876 23.1254 25.2501 23.1254H23.0001V11.4379H24.9376ZM13.0001 3.14727L21.472 9.30976H4.52824L13.0001 3.14727ZM5.25011 11.4379H8.90636V23.1254H5.25011V11.4379ZM11.1564 11.4379H14.8126V23.1254H11.1564V11.4379ZM20.7501 23.1254H17.0626V11.4379H20.7501V23.1254Z"
+							fill="#191918"
+							fill-opacity="0.7"
+						/>
+					</svg>
+
+					<div className="ml-3">
+						<div className="flex">
+							<div className=" text-black-500 text-sm">
+								{`${bank.accountNickname} | ${bank.bankName}`}
+							</div>
+							<span className="ml-2">
+								<svg
+									width="16"
+									height="16"
+									viewBox="0 0 16 16"
+									fill="none"
+									xmlns="http://www.w3.org/2000/svg"
+								>
+									<path
+										d="M10.0669 0.870195C9.79776 0.594736 9.47624 0.375856 9.12128 0.226418C8.76632 0.0769807 8.38507 0 7.99993 0C7.61479 0 7.23354 0.0769807 6.87857 0.226418C6.52361 0.375856 6.2021 0.594736 5.93293 0.870195L5.31093 1.5082L4.42093 1.4972C4.03568 1.49264 3.6534 1.56515 3.29659 1.71048C2.93977 1.85581 2.61562 2.07102 2.34319 2.34345C2.07076 2.61589 1.85555 2.94004 1.71022 3.29686C1.56489 3.65367 1.49237 4.03595 1.49693 4.4212L1.50693 5.3112L0.870928 5.9332C0.595468 6.20237 0.376588 6.52388 0.227151 6.87884C0.0777131 7.23381 0.000732422 7.61506 0.000732422 8.0002C0.000732422 8.38533 0.0777131 8.76659 0.227151 9.12155C0.376588 9.47651 0.595468 9.79803 0.870928 10.0672L1.50793 10.6892L1.49693 11.5792C1.49237 11.9644 1.56489 12.3467 1.71022 12.7035C1.85555 13.0603 2.07076 13.3845 2.34319 13.6569C2.61562 13.9294 2.93977 14.1446 3.29659 14.2899C3.6534 14.4352 4.03568 14.5078 4.42093 14.5032L5.31093 14.4932L5.93293 15.1292C6.2021 15.4047 6.52361 15.6235 6.87857 15.773C7.23354 15.9224 7.61479 15.9994 7.99993 15.9994C8.38507 15.9994 8.76632 15.9224 9.12128 15.773C9.47624 15.6235 9.79776 15.4047 10.0669 15.1292L10.6889 14.4922L11.5789 14.5032C11.9642 14.5078 12.3465 14.4352 12.7033 14.2899C13.0601 14.1446 13.3842 13.9294 13.6567 13.6569C13.9291 13.3845 14.1443 13.0603 14.2896 12.7035C14.435 12.3467 14.5075 11.9644 14.5029 11.5792L14.4929 10.6892L15.1289 10.0672C15.4044 9.79803 15.6233 9.47651 15.7727 9.12155C15.9221 8.76659 15.9991 8.38533 15.9991 8.0002C15.9991 7.61506 15.9221 7.23381 15.7727 6.87884C15.6233 6.52388 15.4044 6.20237 15.1289 5.9332L14.4919 5.3112L14.5029 4.4212C14.5075 4.03595 14.435 3.65367 14.2896 3.29686C14.1443 2.94004 13.9291 2.61589 13.6567 2.34345C13.3842 2.07102 13.0601 1.85581 12.7033 1.71048C12.3465 1.56515 11.9642 1.49264 11.5789 1.4972L10.6889 1.5072L10.0669 0.871195V0.870195ZM10.3539 6.8542L7.35393 9.8542C7.30748 9.90076 7.25231 9.9377 7.19156 9.96291C7.13082 9.98811 7.06569 10.0011 6.99993 10.0011C6.93416 10.0011 6.86904 9.98811 6.80829 9.96291C6.74755 9.9377 6.69237 9.90076 6.64593 9.8542L5.14593 8.3542C5.09944 8.30771 5.06256 8.25252 5.0374 8.19178C5.01225 8.13104 4.9993 8.06594 4.9993 8.0002C4.9993 7.93445 5.01225 7.86935 5.0374 7.80861C5.06256 7.74787 5.09944 7.69268 5.14593 7.6462C5.19242 7.59971 5.2476 7.56283 5.30834 7.53767C5.36908 7.51251 5.43418 7.49956 5.49993 7.49956C5.56567 7.49956 5.63077 7.51251 5.69151 7.53767C5.75225 7.56283 5.80744 7.59971 5.85393 7.6462L6.99993 8.7932L9.64593 6.1462C9.73981 6.05231 9.86715 5.99956 9.99993 5.99956C10.1327 5.99956 10.26 6.05231 10.3539 6.1462C10.4478 6.24008 10.5006 6.36742 10.5006 6.5002C10.5006 6.63297 10.4478 6.76031 10.3539 6.8542Z"
+										fill="#64C093"
+									/>
+								</svg>
+							</span>
+						</div>
+						<div className=" text-gray-600 text-sm">
+							************
+							{bank.lastFourDigits}
+						</div>
+					</div>
+				</div>
+				<div className="flex">
+					<div className="rounded-lg bg-gray-200 p-2">
+						<span className="text-green-500 text-sm">Primary</span>
+					</div>
+					<button
+						type="button"
+						className="bg-white hover:bg-gray-200 font-bold py-2 px-2 rounded text-sm"
+					>
+						Make Primary
+					</button>
+					<button
+						type="button"
+						className="bg-white hover:bg-gray-200 font-bold py-2 px-2 rounded text-sm"
+					>
+						<svg
+							width="16"
+							height="18"
+							viewBox="0 0 16 18"
+							fill="none"
+							xmlns="http://www.w3.org/2000/svg"
+						>
+							<path
+								d="M0.5 2.75V4H1.75V16.5C1.75 16.8315 1.8817 17.1495 2.11612 17.3839C2.35054 17.6183 2.66848 17.75 3 17.75H13C13.3315 17.75 13.6495 17.6183 13.8839 17.3839C14.1183 17.1495 14.25 16.8315 14.25 16.5V4H15.5V2.75H0.5ZM3 16.5V4H13V16.5H3ZM5.5 0.25H10.5V1.5H5.5V0.25Z"
+								fill="#191918"
+								fill-opacity="0.7"
+							/>
+						</svg>
+					</button>
+				</div>
+			</div>
+		</>
+	);
+};
+
+export default BankAccountListItem;
